Building a little more of the server side stuff.
This is still just placeholder stuff. The initial list it will return wi
ll still just be a canned list of comics to test out the client side som
e more. But soon I will try to have it pull a set of comics from a direc
tory, take them apart and add them to some kind of server side data stru
ctures that it can share with the client. It won't be much but it'll be
a start.