The list of comics for browsing now comes from the server as well.
Even though the server in this version is still serving up hard coded da
ta, the client (though extremely simple) is adequate to let you choose a
nd read from a set of comics the server can serve up. So it's time to de
vote more attention to the server.